requirements.txt
# requirements.txt

langchain>=0.3.0
langgraph>=0.1.1
langchain-google-genai>=1.0.8
pydantic>=2.7.0
streamlit>=1.35.0
python-dotenv>=1.0.1
rich>=13.7.0

# 非同期イベントループのネストを許可するために必要
nest-asyncio>=1.6.0